requirements.txt
pandas==2.2.3
typing==3.7.4.3
openai==1.76.0
SQLAlchemy==2.0.40
python-dotenv==1.1.0
smolagents
pydantic==2.7.1
pytest==7.4.0
